Handle new RID mapping TSV file. Add message for mapped RIDs.
Handle a new TSV file for RID mapping.

The name of the new TSV file appears as the last column in the main trunking config file. trunking.py will read this new TSV file and store it in `srcid_map`.

trunking.py now sends a new `src_tag` message. When a new transmission is received with an RID, trunking.py will attempt to look it up in srcid_map. If there is an entry for it, the src_tag message will have the string that was in the second column of the TSV file. If there is not an entry for it, the src_tag message will have the RID digits, only, as a string.

The new TSV file for RID mapping has two columns. The first column is the RID, and the second column is a text description of that RID. There should be one row per RID. It should look something like this:

1234\tMetropolis Fire Engine 3
4567\tMetropolis Public Works Plow 1
101112\tGotham Ambulance Dispatch

The new TSV file does not support ranges - you can't say something like `1000-1300\tMetropolis Fire`.

With the above example, if a transmission is received from RID 1234, the `src_tag` message would be `Metropolis Fire Engine 3`. If a transmission came in from RID 131415, the `src_tag` message would be `131415`.