Don't send get_terminal_config message on startup.
Don't send the get_terminal_config message on startup. I am not sure why this is needed, but I have observed that if that message is sent, OP25 crashes shortly after starting.

I *think* it might be some kind of race condition, because trunking.py now has to parse the new TSV file with the RID mapping. In my configuration, the new TSV file has 2500+ RID mappings in it; the extra time needed to parse that file may be causing a problem.