SmartNet: remove default timestamp parameter
When you do a default parameter, it is executed once, at the beginning of the program. That means that `ts=time.time()` sets `ts` to the start time of the program, rather than the call time of the function. At best this is wrong, at worst it confuses the programmer (me). Remove the default parameter since it's not used anyway.