Web UI: increase width from 730px to 830px
It's 2024, not 2020 (de679e9cd6f7afd96702e426c9ba86ba29d064e6) - we can afford 100 more pixels of width and still display two windows side by side on even a minimal HD 1080p monitor.

In all seriousness, though, this is needed to display call mode for SmartNet, which is valuable enough to be worth it.